Maven: Update exceed dependency
Use current 0.3.2 build of exceed.

Update the frontend-maven-plugin execution to the current versions of
plugin, nodejs and yarn, but comment it out for now since we have no
app js build so far.

Configure an execution of the GenerateResourceLists helper to generate
resource.lst files for all extensions in the project. (WIP)