Change DSLContext injection
With the new Atomikos data source setup we can no longer rely on spring
to provide us with a DSLContext because there is no longer the one
global DSLContext/dats source

For action components we now have a ParamterProvider for DSLContexts
which we use in ShopActions and also in the CreateOrdersTest "test".

We still need an easy way to get a DataSource/DSLContext in a test.